db: refactored uow interface
